+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - opening motion photo embedded video when video track is not the first one
+- some SVG rendering issues
 
 ## <a id="v1.12.3"></a>[v1.12.3] - 2025-02-06